Proof of stake code to miner (#34)
* Add POS shell to CreateNewBlock

+++ src/accumulators.h
@@ -25,7 +25,7 @@ void DatabaseChecksums(AccumulatorMap& mapAccumulators);
 bool LoadAccumulatorValuesFromDB(const uint256 nCheckpoint);
 bool EraseAccumulatorValues(const uint256& nCheckpointErase, const uint256& nCheckpointPrevious);
 uint256 GetChecksum(const CBigNum &bnValue);
-int GetChecksumHeight(uint32_t nChecksum, libzerocoin::CoinDenomination denomination);
+int GetChecksumHeight(uint256 nChecksum, libzerocoin::CoinDenomination denomination);
 bool InvalidCheckpointRange(int nHeight);
 bool ValidateAccumulatorCheckpoint(const CBlock& block, CBlockIndex* pindex, AccumulatorMap& mapAccumulators);
